feat/#61: log 테이블 수정 및 추가
