--- src/main/java/com/epi/epilog/app/domain/diabets/LogExercise.java
@@ -1,18 +1,0 @@
-package com.epi.epilog.app.domain.diabets;
-
-import jakarta.persistence.Entity;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import static lombok.AccessLevel.PRIVATE;
-import static lombok.AccessLevel.PROTECTED;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = PROTECTED)
-@AllArgsConstructor(access = PRIVATE)
-public class LogExercise {
-}
