+++ src/main/java/com/epi/epilog/app/domain/log/LogMood.java
@@ -1,0 +1,32 @@
+package com.epi.epilog.app.domain.log;
+
+import io.micrometer.common.lang.Nullable;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.ColumnDefault;
+
+import static lombok.AccessLevel.PRIVATE;
+import static lombok.AccessLevel.PROTECTED;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = PROTECTED)
+@AllArgsConstructor(access = PRIVATE)
+public class LogMood {
+    @Id
+    @Column(name="log_mood_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="log_id")
+    private Log log;
+    @ColumnDefault("false")
+    private boolean detailsState;
+    private String type;
+    @Nullable
+    private String details;
+}
